+++ subprojects/gdk_rust/gdk_common/src/descriptor.rs
@@ -1,7 +1,15 @@
 use crate::error::Error;
 use crate::scripts::ScriptType;
+use crate::wally::MasterBlindingKey;
 use bitcoin::bip32::{ChildNumber, Fingerprint, Xpub};
+use elements_miniscript::confidential::Key;
+use elements_miniscript::{
+    ConfidentialDescriptor, DescriptorPublicKey as ElementsDescriptorPublicKey,
+};
+use miniscript::descriptor::checksum::desc_checksum;
 use miniscript::descriptor::{Descriptor, DescriptorPublicKey, ShInner};
+use std::convert::TryInto;
+use std::str::FromStr;
 
 /// Make sure the key origin is in the expected format
 /// and return the bip32 account number
@@ -29,11 +37,12 @@ fn check_xpub_consitency(
     xpub: Xpub,
     bip32_account: u32,
     fingerprint: Fingerprint,
-) -> Result<(ScriptType, Xpub, u32, Fingerprint), Error> {
+    master_blinding_key: Option<MasterBlindingKey>,
+) -> Result<(ScriptType, Xpub, u32, Fingerprint, Option<MasterBlindingKey>), Error> {
     match xpub.child_number {
         ChildNumber::Hardened {
             index: n,
-        } if n == bip32_account => Ok((script_type, xpub, n, fingerprint)),
+        } if n == bip32_account => Ok((script_type, xpub, n, fingerprint, master_blinding_key)),
         _ => Err(Error::UnsupportedDescriptor),
     }
 }
@@ -42,9 +51,30 @@ fn check_xpub_consitency(
 pub fn parse_single_sig_descriptor(
     s: &str,
     coin_type: u32,
-) -> Result<(ScriptType, Xpub, u32, Fingerprint), Error> {
+    is_liquid: bool,
+) -> Result<(ScriptType, Xpub, u32, Fingerprint, Option<MasterBlindingKey>), Error> {
+    let (mbk, s) = if is_liquid {
+        let descriptor = ConfidentialDescriptor::<ElementsDescriptorPublicKey>::from_str(s)
+            .map_err(|_| Error::UnsupportedDescriptor)?;
+        let mbk = match descriptor.key {
+            Key::Slip77(mbk) => {
+                let mut v = mbk.as_bytes().to_vec();
+                // mbk is 32 bytes, make it 64 as wally expects
+                v.splice(0..0, [0u8; 32]);
+                MasterBlindingKey(v.try_into().map_err(|_| Error::UnsupportedDescriptor)?)
+            }
+            _ => return Err(Error::UnsupportedDescriptor),
+        };
+        // remove "el" and elements checksum, then compute the bitcoin checksum
+        let desc = descriptor.descriptor.to_string();
+        let btc_desc = &desc[2..(desc.len() - 9)];
+        let checksum = desc_checksum(&btc_desc).map_err(|_| Error::UnsupportedDescriptor)?;
+        (Some(mbk), format!("{}#{}", &btc_desc, checksum))
+    } else {
+        (None, s.to_string())
+    };
     let (desc, _) =
-        Descriptor::parse_descriptor(&crate::EC, s).map_err(|_| Error::UnsupportedDescriptor)?;
+        Descriptor::parse_descriptor(&crate::EC, &s).map_err(|_| Error::UnsupportedDescriptor)?;
     if !desc.has_wildcard() {
         return Err(Error::UnsupportedDescriptor);
     }
@@ -59,6 +89,7 @@ pub fn parse_single_sig_descriptor(
                         descriptorxkey.xkey,
                         n,
                         *f,
+                        mbk,
                     );
                 }
             }
@@ -67,14 +98,14 @@ pub fn parse_single_sig_descriptor(
         if let DescriptorPublicKey::XPub(descriptorxkey) = wpkh.as_inner() {
             if let Some((f, p)) = &descriptorxkey.origin {
                 let n = match_key_origin(&p.clone().into(), 84, coin_type)?;
-                return check_xpub_consitency(ScriptType::P2wpkh, descriptorxkey.xkey, n, *f);
+                return check_xpub_consitency(ScriptType::P2wpkh, descriptorxkey.xkey, n, *f, mbk);
             }
         }
     } else if let Descriptor::Pkh(pkh) = desc {
         if let DescriptorPublicKey::XPub(descriptorxkey) = pkh.as_inner() {
             if let Some((f, p)) = &descriptorxkey.origin {
                 let n = match_key_origin(&p.clone().into(), 44, coin_type)?;
-                return check_xpub_consitency(ScriptType::P2pkh, descriptorxkey.xkey, n, *f);
+                return check_xpub_consitency(ScriptType::P2pkh, descriptorxkey.xkey, n, *f, mbk);
             }
         }
     }
@@ -84,6 +115,7 @@ pub fn parse_single_sig_descriptor(
 #[cfg(test)]
 mod test {
     use super::*;
+    use elements::hex::ToHex;
 
     #[test]
     fn test_descriptor() {
@@ -102,36 +134,42 @@ mod test {
         let p2wpkh_incorrect_key_origin1 = format!("sh(wpkh([00000000/44'/1'/0']{}/0/*))", tpub);
         let p2wpkh_incorrect_key_origin2 = format!("sh(wpkh([00000000/84'/1'/0'/0']{}/0/*))", tpub);
 
+        let is_liquid = false;
         // Valid cases
-        let (t, shp2wpkh_xpub_external, bip32_account, f) =
-            parse_single_sig_descriptor(&shp2wpkh, coin_type).unwrap();
+        let (t, shp2wpkh_xpub_external, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&shp2wpkh, coin_type, is_liquid).unwrap();
         assert_eq!(t, ScriptType::P2shP2wpkh);
         assert_eq!(bip32_account, 0);
         assert_eq!(f, Fingerprint::default());
-        let (t, shp2wpkh_xpub_internal, bip32_account, f) =
-            parse_single_sig_descriptor(&shp2wpkh_change, coin_type).unwrap();
+        assert!(mbk.is_none());
+        let (t, shp2wpkh_xpub_internal, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&shp2wpkh_change, coin_type, is_liquid).unwrap();
         assert_eq!(t, ScriptType::P2shP2wpkh);
         assert_eq!(bip32_account, 0);
         assert_eq!(f, Fingerprint::default());
-        let (t, p2wpkh_xpub, bip32_account, f) =
-            parse_single_sig_descriptor(&p2wpkh, coin_type).unwrap();
+        assert!(mbk.is_none());
+        let (t, p2wpkh_xpub, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&p2wpkh, coin_type, is_liquid).unwrap();
         assert_eq!(t, ScriptType::P2wpkh);
         assert_eq!(bip32_account, 0);
         assert_eq!(f, Fingerprint::default());
-        let (t, p2wpkh_xpub_1, bip32_account, f) =
-            parse_single_sig_descriptor(&p2wpkh_1, coin_type).unwrap();
+        assert!(mbk.is_none());
+        let (t, p2wpkh_xpub_1, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&p2wpkh_1, coin_type, is_liquid).unwrap();
         assert_eq!(t, ScriptType::P2wpkh);
         assert_eq!(bip32_account, 1);
         assert_eq!(f, Fingerprint::default());
-        let (t, p2pkh_xpub, bip32_account, f) =
-            parse_single_sig_descriptor(&p2pkh, coin_type).unwrap();
+        assert!(mbk.is_none());
+        let (t, p2pkh_xpub, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&p2pkh, coin_type, is_liquid).unwrap();
         assert_eq!(t, ScriptType::P2pkh);
         assert_eq!(bip32_account, 0);
         assert_eq!(f, Fingerprint::default());
+        assert!(mbk.is_none());
 
         // Invalid cases
         let err_str = Error::UnsupportedDescriptor.to_string();
-        let f = |(s, t)| parse_single_sig_descriptor(s, t).unwrap_err().to_string();
+        let f = |(s, t)| parse_single_sig_descriptor(s, t, is_liquid).unwrap_err().to_string();
         assert_eq!(f((tpub, coin_type)), err_str);
         assert_eq!(f((tpub, 0)), err_str);
         assert_eq!(f((&shmulti, coin_type)), err_str);
@@ -147,5 +185,24 @@ mod test {
         assert_eq!(p2wpkh_xpub.to_string(), tpub);
         assert_eq!(p2wpkh_xpub_1.to_string(), tpub_1);
         assert_eq!(p2pkh_xpub.to_string(), tpub);
+
+        // Liquid test cases
+        let is_liquid = true;
+
+        let slip77_key = "0000000000000000000000000000000000000000000000000000000000000011";
+        let shp2wpkh_slip77 = format!("ct(slip77({}),el{})", slip77_key, shp2wpkh);
+        let (t, shp2wpkh_xpub_external, bip32_account, f, mbk) =
+            parse_single_sig_descriptor(&shp2wpkh_slip77, coin_type, is_liquid).unwrap();
+        assert_eq!(t, ScriptType::P2shP2wpkh);
+        assert_eq!(shp2wpkh_xpub_external.to_string(), tpub);
+        assert_eq!(bip32_account, 0);
+        assert_eq!(f, Fingerprint::default());
+        assert_eq!(slip77_key, mbk.unwrap().0[32..].to_hex());
+
+        // View key not supported
+        let view_key = "cU5b4e1Zv5Kgda8A2Pn8SQBKAGq6PNZbzk14eHrFpHby7c8xKjLP";
+        let shp2wpkh_view = format!("ct({},el{})", view_key, shp2wpkh);
+        let f = |(s, t)| parse_single_sig_descriptor(s, t, is_liquid).unwrap_err().to_string();
+        assert_eq!(f((&shp2wpkh_view, coin_type)), err_str);
     }
 }